# packages/shared/openapi/backend-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Backend.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/profiles": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "operationId": "listProfiles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Profiles"
        ],
        "operationId": "createProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/{profileId}": {
      "put": {
        "tags": [
          "Profiles"
        ],
        "operationId": "updateProfile",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profiles"
        ],
        "operationId": "deleteProfile",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/library/{kind}": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "listLibrary",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryPage"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/library/{kind}/{masterId}": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "getLibraryItem",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "masterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "LibraryPage": {
        "required": [
          "total",
          "items"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MasterCard"
            }
          }
        }
      },
...
      "MasterCard": {
        "required": [
          "id",
          "title",
          "year",
          "posterUrl",
          "rating",
          "bestQuality",
          "variantCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "posterUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "rating": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "bestQuality": {
            "type": [
              "null",
              "string"
            ]
          },
          "variantCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "MasterDetails": {
        "required": [
          "id",
          "title",
          "year",
          "posterUrl",
          "rating",
          "bestQuality",
          "variants"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "posterUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "rating": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "bestQuality": {
            "type": [
              "null",
              "string"
            ]
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantInfo"
            }
          }
        }
      },
...
      "ProblemDetails": {
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProfileDto": {
        "required": [
          "id",
          "name",
          "avatarKey",
          "isKids"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "avatarKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "isKids": {
            "type": "boolean"
          }
        }
      },
      "ProfileRequest": {
        "required": [
          "name",
          "isKids"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "avatarKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "isKids": {
            "type": "boolean"
          }
        }
      },
...
      "VariantInfo": {
        "required": [
          "streamId",
          "label",
          "quality",
          "source",
          "audioLanguages",
          "audioTag",
          "isHdr",
          "containerExtension",
          "categoryId",
          "rawTitle"
        ],
        "type": "object",
        "properties": {
          "streamId": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "quality": {
            "type": [
              "null",
              "string"
            ]
          },
          "source": {
            "type": [
              "null",
              "string"
            ]
          },
          "audioLanguages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "audioTag": {
            "type": [
              "null",
              "string"
            ]
          },
          "isHdr": {
            "type": "boolean"
          },
          "containerExtension": {
            "type": [
              "null",
              "string"
            ]
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ]
          },
          "rawTitle": {
            "type": "string"
          }
        }
      }
    }
  },
...
}